// migrations/snapshots/757f866c81994f7be445a65ba3c2b5543720a98e7ab645e871b230a876c46318/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly otp: CodecTypes['pg/text@1']['output'];
    };
    readonly IdentityType: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Skill: {
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly SkillCategory: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly sortOrder: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/bool@1']['output'];
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly UserCuriosity: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly skillId: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly UserIdentity: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly identityTypeId: CodecTypes['pg/int4@1']['output'];
      readonly isPrimary: CodecTypes['pg/bool@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly UserProfile: {
      readonly ageRange: CodecTypes['pg/text@1']['output'];
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly countryCode: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
    readonly UserSkill: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly relationshipType: 'CAN_DO' | 'LEARNING' | 'WANT_TO_IMPROVE';
      readonly skillId: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly otp: CodecTypes['pg/text@1']['input'];
    };
    readonly IdentityType: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Skill: {
      readonly categoryId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly SkillCategory: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly sortOrder: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/bool@1']['input'];
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly UserCuriosity: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly skillId: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly UserIdentity: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly identityTypeId: CodecTypes['pg/int4@1']['input'];
      readonly isPrimary: CodecTypes['pg/bool@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly UserProfile: {
      readonly ageRange: CodecTypes['pg/text@1']['input'];
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly countryCode: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
    readonly UserSkill: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly relationshipType: 'CAN_DO' | 'LEARNING' | 'WANT_TO_IMPROVE';
      readonly skillId: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;