// migrations/snapshots/c6135dcb8b48032f1f0973789594a9809a6cb0cf81455201357208400991ba55/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly otp: CodecTypes['pg/text@1']['output'];
    };
    readonly EvolvProfile: {
      readonly ageRange: 'UNDER_18' | 'AGE_18_24' | 'AGE_25_34' | 'AGE_35_44' | 'AGE_45_PLUS';
      readonly completedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly country: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly curiositySkills: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly focusAreas: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly focusSubAreas: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly profileCompleted: CodecTypes['pg/bool@1']['output'];
      readonly profilePhoto: CodecTypes['pg/text@1']['output'] | null;
      readonly roles: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsCanDo: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsLearning: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsToImprove: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly workingToward: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/bool@1']['output'];
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly otp: CodecTypes['pg/text@1']['input'];
    };
    readonly EvolvProfile: {
      readonly ageRange: 'UNDER_18' | 'AGE_18_24' | 'AGE_25_34' | 'AGE_35_44' | 'AGE_45_PLUS';
      readonly completedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly country: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly curiositySkills: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly focusAreas: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly focusSubAreas: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly profileCompleted: CodecTypes['pg/bool@1']['input'];
      readonly profilePhoto: CodecTypes['pg/text@1']['input'] | null;
      readonly roles: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsCanDo: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsLearning: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsToImprove: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly workingToward: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/bool@1']['input'];
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly otp: CodecTypes['pg/text@1']['output'];
    };
    readonly EvolvProfile: {
      readonly ageRange: 'UNDER_18' | 'AGE_18_24' | 'AGE_25_34' | 'AGE_35_44' | 'AGE_45_PLUS';
      readonly completedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly country: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly curiositySkills: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly focusAreas: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly focusSubAreas: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly profileCompleted: CodecTypes['pg/bool@1']['output'];
      readonly profilePhoto: CodecTypes['pg/text@1']['output'] | null;
      readonly roles: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsCanDo: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsLearning: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly skillsToImprove: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly workingToward: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/bool@1']['output'];
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly EmailOtp: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly otp: CodecTypes['pg/text@1']['input'];
    };
    readonly EvolvProfile: {
      readonly ageRange: 'UNDER_18' | 'AGE_18_24' | 'AGE_25_34' | 'AGE_35_44' | 'AGE_45_PLUS';
      readonly completedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly country: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly curiositySkills: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly focusAreas: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly focusSubAreas: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly profileCompleted: CodecTypes['pg/bool@1']['input'];
      readonly profilePhoto: CodecTypes['pg/text@1']['input'] | null;
      readonly roles: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsCanDo: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsLearning: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly skillsToImprove: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly workingToward: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/bool@1']['input'];
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;